added ui

+++ app/models.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, String, Date, Boolean, ForeignKey, DECIMAL, JSON
+from sqlalchemy import Column, BigInteger, Integer, String, Date, Boolean, ForeignKey, DECIMAL, JSON
 from .database import Base
 
 class Driver(Base):
     __tablename__ = "driver"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     number = Column(Integer)
     name = Column(String(100))
     nationality = Column(String(50))
@@ -15,7 +15,7 @@ class Driver(Base):
 class Circuit(Base):
     __tablename__ = "circuit"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String(100))
     location = Column(String(100))
     length = Column(DECIMAL(10, 2))
@@ -25,8 +25,8 @@ class Circuit(Base):
 class Race(Base):
     __tablename__ = "race"
 
-    driver_id = Column(Integer, ForeignKey("driver.id"), primary_key=True)
-    circuit_id = Column(Integer, ForeignKey("circuit.id"), primary_key=True)
+    driver_id = Column(BigInteger, ForeignKey("driver.id"), primary_key=True)
+    circuit_id = Column(BigInteger, ForeignKey("circuit.id"), primary_key=True)
     race_date = Column(Date, primary_key=True)
     place = Column(Integer)
     points = Column(Integer)
